% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{autocomplete}
\alias{autocomplete}
\title{This function autocompletes by determining whether to  autocomplete the last set of characters or  predict a new word based on
previous words. The determining factor is the ending whitespace. Including an ending whitespace to signify new word, else
the predictor predicts the next word based on the last set of charactes in the string}
\usage{
autocomplete(str)
}
\arguments{
\item{str}{a string that contains the words to be used to predict next word}
}
\value{
a data.frame containing the top 3 most likely words and their respective log probabilities
}
\description{
This function autocompletes by determining whether to  autocomplete the last set of characters or  predict a new word based on
previous words. The determining factor is the ending whitespace. Including an ending whitespace to signify new word, else
the predictor predicts the next word based on the last set of charactes in the string
}
\details{
End with a space if you hope to predict a new word based on previous words, else it will try to autocomplete the
last word as if was the beginning of an unknown predicted word
}
\author{
Jerome Cordjotse
}
