% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{autocomplete_checkingSpace}
\alias{autocomplete_checkingSpace}
\title{Building an autocopleting fucntion to autocomplete or predict next word}
\usage{
autocomplete_checkingSpace(str, spaced = T)
}
\arguments{
\item{str}{a string that contains the words to be used to predict next word}

\item{spaced}{a logical that determines where to complete the last set or characters or predict a new word}
}
\value{
a data.frame containing the top 3 most likely words and their respective log probabilities
}
\description{
This function autocompletes whether to autocomplete the last set of characters or  predict a new word based on
previous words determined by the spaced variable. Spaced is true if you want a new word but to autocomplete based on
set of characters spaced is false
}
\details{
This function helps if you have an external trigger to determine whether to complete the word from the last
set of characters or predict a new word based on previous words
}
\author{
Jerome Cordjotse
}
