% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{quadgram}
\alias{quadgram}
\title{Quadgram model}
\format{
A data frame with 1million plus rows and 6 variables:
\describe{
\item{ngram.1}{quadgrams, 1st word for the set of words}
\item{ngram.2}{quadgrams, 2nd word for the set of words}
\item{ngram.3}{quadgrams, 3rd word for the set of words}
\item{ngram.4}{quadgrams, 4th word for the set of words}
\item{prob}{probability of the ngram}
...
}
}
\source{
processed loacally by jerome
}
\usage{
quadgram
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
