% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{trigram}
\alias{trigram}
\title{Trigram model}
\format{
A data frame with 1million plus rows and 5 variables:
\describe{
\item{ngram.1}{trigrams, 1st word for the set of words}
\item{ngram.2}{trigrams, 2nd word for the set of words}
\item{ngram.3}{trigrams, 3rd word for the set of words}
\item{prob}{probability of the ngram}
...
}
}
\source{
processed loacally by jerome
}
\usage{
trigram
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
