% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{unigram}
\alias{unigram}
\title{Unigram model}
\format{
A data frame with 1million plus rows and 3 variables:
\describe{
\item{ngram}{unigrams, for the set of words}
\item{prob}{probability of the ngram}
...
}
}
\source{
processed loacally by jerome
}
\usage{
unigram
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
